requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
mysqlclient==2.2.4
pymysql==1.1.1
pydantic==2.8.2
passlib==1.7.4
bcrypt==4.2.0
python-jose==3.3.0
boto3==1.34.145
pytz==2024.1
redis==5.0.8
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
oauthlib==3.2.2
httpx==0.27.2